# api/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Comprehensive Electronic Medical Record system for hospitals",
        "title": "Hospital EMR System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@hospital-emr.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "http://www.hospital-emr.com/license"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/appointments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a paginated list of appointments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "appointments"
                ],
                "summary": "List appointments",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by patient ID",
                        "name": "patient_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by provider ID",
                        "name": "provider_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by date (YYYY-MM-DD)",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Schedule a new appointment for a patient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "appointments"
                ],
                "summary": "Create a new appointment",
                "parameters": [
                    {
                        "description": "Appointment information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_scheduling.CreateAppointmentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Appointment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_common_errors.AppError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_common_errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/v1/appointments/availability": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get available time slots for a provider (alias for GetProviderAvailability)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "appointments"
                ],
                "summary": "Get general availability",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Provider ID",
                        "name": "provider_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Date (YYYY-MM-DD)",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_scheduling.TimeSlot"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_common_errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/v1/appointments/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed information about an appointment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "appointments"
                ],
                "summary": "Get appointment by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Appointment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Appointment"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_common_errors.AppError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update appointment information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "appointments"
                ],
                "summary": "Update appointment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Appointment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Appointment information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_scheduling.CreateAppointmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Appointment"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_common_errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/v1/appointments/{id}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancel an appointment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "appointments"
                ],
                "summary": "Cancel appointment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Appointment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Cancellation reason",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_common_errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/v1/appointments/{id}/checkin": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mark appointment as checked in",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "appointments"
                ],
                "summary": "Check in appointment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Appointment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_common_errors.AppError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/providers/{id}/availability": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get available time slots for a provider on a specific date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "appointments"
                ],
                "summary": "Get provider availability",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Provider ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Date (YYYY-MM-DD)",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_scheduling.TimeSlot"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_common_errors.AppError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_hospital-emr_backend_internal_common_errors.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_hospital-emr_backend_internal_models.Appointment": {
            "type": "object",
            "properties": {
                "appointment_number": {
                    "type": "string"
                },
                "appointment_type": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.AppointmentType"
                },
                "cancellation_reason": {
                    "type": "string"
                },
                "cancelled_at": {
                    "type": "string"
                },
                "checked_in_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "department": {
                    "type": "string"
                },
                "duration": {
                    "description": "in minutes",
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "patient": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Patient"
                },
                "patient_id": {
                    "type": "string"
                },
                "provider": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.User"
                },
                "provider_id": {
                    "type": "string"
                },
                "reason_for_visit": {
                    "type": "string"
                },
                "reminder_sent": {
                    "type": "boolean"
                },
                "reminder_sent_at": {
                    "type": "string"
                },
                "room": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.AppointmentStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
        "github_com_hospital-emr_backend_internal_models.AppointmentStatus": {
            "type": "string",
            "enum": [
                "scheduled",
                "confirmed",
                "checked_in",
                "in_progress",
                "completed",
                "cancelled",
                "no_show"
            ],
            "x-enum-varnames": [
                "AppointmentStatusScheduled",
                "AppointmentStatusConfirmed",
                "AppointmentStatusCheckedIn",
                "AppointmentStatusInProgress",
                "AppointmentStatusCompleted",
                "AppointmentStatusCancelled",
                "AppointmentStatusNoShow"
            ]
        },
        "github_com_hospital-emr_backend_internal_models.AppointmentType": {
            "type": "string",
            "enum": [
                "consultation",
                "follow_up",
                "wellness",
                "procedure",
                "emergency",
                "telehealth"
            ],
            "x-enum-varnames": [
                "AppointmentTypeConsultation",
                "AppointmentTypeFollowUp",
                "AppointmentTypeWellness",
                "AppointmentTypeProcedure",
                "AppointmentTypeEmergency",
                "AppointmentTypeTelehealth"
            ]
        },
...
        "github_com_hospital-emr_backend_internal_models.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "department": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_login_at": {
                    "type": "string"
                },
                "last_login_ip": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "license_number": {
                    "type": "string"
                },
                "mfa_enabled": {
                    "type": "boolean"
                },
                "password_expiry": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Role"
                    }
                },
                "specialty": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.UserStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_hospital-emr_backend_internal_models.UserStatus": {
            "type": "string",
            "enum": [
                "active",
                "inactive",
                "suspended",
                "locked"
            ],
            "x-enum-varnames": [
                "UserStatusActive",
                "UserStatusInactive",
                "UserStatusSuspended",
                "UserStatusLocked"
            ]
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "internal_scheduling.CreateAppointmentRequest": {
            "type": "object",
            "required": [
                "appointment_type",
                "duration",
                "patient_id",
                "provider_id",
                "start_time"
            ],
            "properties": {
                "appointment_type": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.AppointmentType"
                },
                "department": {
                    "type": "string"
                },
                "duration": {
                    "description": "in minutes",
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "patient_id": {
                    "type": "string"
                },
                "provider_id": {
                    "type": "string"
                },
                "reason_for_visit": {
                    "type": "string"
                },
                "room": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "internal_scheduling.TimeSlot": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                },
                "end_time": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}